pyproject.toml
[tool.poetry]
name = "power_systems_data_api_demonstrator"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = { version = "^0.20.0", extras = ["standard"] }
pyyaml = "^6.0"
sqlmodel = { git = "https://github.com/AntonDeMeester/sqlmodel.git", rev="main" }
fastapi = "^0.103.2"
pydantic-settings = "^2.0.3"
yarl = "^1.9.2"
aiosqlite = "^0.19.0"
ujson = "^5.8.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
flake8 = "~4.0.1"
mypy = "^1.1.1"
isort = "^5.11.4"
pre-commit = "^3.0.1"
wemake-python-styleguide = "^0.17.0"
black = "^22.12.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.2"
pytest-env = "^0.8.1"
fakeredis = "^2.5.0"
httpx = "^0.23.3"
pandas = "^1.5.3"
jupyter = "^1.0.0"
requests = "^2.28.2"
pytest-asyncio = "^0.21.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["power_systems_data_api_demonstrator",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "POWER_SYSTEMS_DATA_API_DEMONSTRATOR_DB_FILE=test_db.sqlite3",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
